requirements.txt
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
dlib==19.24.6
Flask==2.1.1
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
moviepy==1.0.3
mutagen==1.47.0
numpy==1.22.4
opencv-python==4.6.0.66
Pillow==9.1.1
proglog==0.1.10
pycryptodomex==3.20.0
pytube==15.0.0
pytube3==9.6.4
requests==2.32.3
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
Werkzeug==2.2.3
yt-dlp==2024.8.6
zipp==3.19.2
